Nest pm manager
